fix: get create_signed_url working
